Move icon component to correct folder
